requirements.txt
pandas
polars>0.20.0
pyarrow
tabulate
pysftp
synapseclient
ipykernel
python-dotenv
pip
mne
seaborn
invoke
nbconvert
great_tables
autoreject
mne_bids
pybv
eeglabio
EDFlib-Python
mne-bids-pipeline
ipyevents
jupyterlab>=4.0.0
jupytext
mne-qt-browser
PyQt6
neurodsp>=2.3
specparam>=2.0.0rc1
selenium


